refactor: schema loading
